Migrate const length and range analyses
