+++ crates/cubecl-opt/src/analyses/mod.rs
@@ -1,5 +1,7 @@
 mod base;
+pub mod const_len;
 pub mod dominance;
+pub mod integer_range;
 pub mod liveness;
 pub mod post_order;
 pub mod writes;
