+++ crates/cubecl-opt/src/passes/mod.rs
@@ -3,10 +3,9 @@ mod composite;
 mod constant_prop;
 mod dead_code;
 mod expression_merge;
-mod in_bounds_analysis;
+mod find_safe_indexes;
 mod index_merge;
 mod inlined_if_to_select;
-mod integer_range_analysis;
 mod reduce_strength;
 
 pub use array_copy_propagate::*;
@@ -14,10 +13,9 @@ pub use composite::*;
 pub use constant_prop::*;
 pub use dead_code::*;
 pub use expression_merge::*;
-pub use in_bounds_analysis::*;
+pub use find_safe_indexes::*;
 pub use index_merge::*;
 pub use inlined_if_to_select::*;
-pub use integer_range_analysis::*;
 pub use reduce_strength::*;
 
 use crate::AtomicCounter;
